+++ saleor/attribute/error_codes.py
@@ -20,3 +20,15 @@ class AttributeBulkCreateErrorCode(Enum):
     UNIQUE = "unique"
     DUPLICATED_INPUT_ITEM = "duplicated_input_item"
     MAX_LENGTH = "max_length"
+
+
+class AttributeBulkUpdateErrorCode(Enum):
+    ALREADY_EXISTS = "already_exists"
+    BLANK = "blank"
+    GRAPHQL_ERROR = "graphql_error"
+    INVALID = "invalid"
+    NOT_FOUND = "not_found"
+    REQUIRED = "required"
+    UNIQUE = "unique"
+    DUPLICATED_INPUT_ITEM = "duplicated_input_item"
+    MAX_LENGTH = "max_length"