+++ saleor/graphql/attribute/mutations/__init__.py
@@ -1,4 +1,5 @@
 from .attribute_bulk_create import AttributeBulkCreate
+from .attribute_bulk_update import AttributeBulkUpdate
 from .attribute_create import AttributeCreate
 from .attribute_delete import AttributeDelete
 from .attribute_reorder_values import AttributeReorderValues
@@ -15,6 +16,7 @@ __all__ = [
     "AttributeCreate",
     "AttributeDelete",
     "AttributeBulkCreate",
+    "AttributeBulkUpdate",
     "AttributeReorderValues",
     "AttributeUpdate",
     "AttributeValueCreate",