+++ saleor/graphql/attribute/schema.py
@@ -17,6 +17,7 @@ from .bulk_mutations import AttributeBulkDelete, AttributeValueBulkDelete
 from .filters import AttributeFilterInput, AttributeWhereInput, filter_attribute_search
 from .mutations import (
     AttributeBulkCreate,
+    AttributeBulkUpdate,
     AttributeCreate,
     AttributeDelete,
     AttributeReorderValues,
@@ -79,6 +80,7 @@ class AttributeMutations(graphene.ObjectType):
     attribute_delete = AttributeDelete.Field()
     attribute_update = AttributeUpdate.Field()
     attribute_bulk_create = AttributeBulkCreate.Field()
+    attribute_bulk_update = AttributeBulkUpdate.Field()
     attribute_translate = AttributeTranslate.Field()
     attribute_bulk_translate = AttributeBulkTranslate.Field()
     attribute_bulk_delete = AttributeBulkDelete.Field()