+++ saleor/graphql/payment/mutations/transaction/transaction_event_report.py
@@ -113,6 +113,7 @@ class TransactionEventReport(ModelMutation):
     def update_transaction(
         cls,
         transaction: payment_models.TransactionItem,
+        transaction_event: payment_models.TransactionEvent,
         available_actions: Optional[list[str]] = None,
         app: Optional["App"] = None,
     ):
@@ -126,6 +127,20 @@ class TransactionEventReport(ModelMutation):
             "refund_pending_value",
             "cancel_pending_value",
         ]
+
+        if (
+            transaction_event.type
+            in [
+                TransactionEventType.AUTHORIZATION_REQUEST,
+                TransactionEventType.AUTHORIZATION_SUCCESS,
+                TransactionEventType.CHARGE_REQUEST,
+                TransactionEventType.CHARGE_SUCCESS,
+            ]
+            and not transaction.psp_reference
+        ):
+            transaction.psp_reference = transaction_event.psp_reference
+            fields_to_update.append("psp_reference")
+
         if available_actions is not None:
             transaction.available_actions = available_actions
             fields_to_update.append("available_actions")
@@ -239,6 +254,7 @@ class TransactionEventReport(ModelMutation):
             previous_refunded_value = transaction.refunded_value
             cls.update_transaction(
                 transaction,
+                transaction_event,
                 available_actions=available_actions,
                 app=app,
             )